bug fixing.. delete old datasets files
