+++ regain/_datasets/ising.py
@@ -23,12 +23,14 @@ def update_ising_l1(theta_init, no, n_dim_obs, responses=[-1,1]):
     return theta
 
 def ising_theta_generator(p=10, n=100, T=10, mode='l1', time_on_axis='first',
-          change=4, responses=[-1,1],
-          update_theta='l1'):
+          change=4, responses=[-1,1]):
     theta = np.random.choice(np.array([-0.5, 0, 0.5]), size=(p,p), replace=True)
     theta += theta.T
     np.fill_diagonal(theta, 0)
+    theta[np.where(theta>0)] = 1
+    theta[np.where(theta<0)] = -1
     thetas = [theta]
+
 
     for t in range(1, T):
         if mode == 'switch':
