+++ regain/datasets.py
@@ -80,22 +80,16 @@ def _gaussian_case(n_samples=100, n_dim_obs=100, n_dim_lat=10, T=10, mode=None,
 
 def _ising_case(n_samples=100, n_dim_obs=100, T=10,
                         time_on_axis='first',update_theta='l2',
-                        responses=[-1,1],
-                        **kwargs):
+                        responses=[-1,1], **kwargs):
         thetas = ising_theta_generator(p=n_dim_obs, n=n_samples, T=T,
                                         mode=update_theta,
                                         **kwargs)
-        data = np.array(
-            [
-                ising_sampler(t,
-                    np.zeros(n_dim_obs), n=n_samples,
-                    responses=responses) for t in thetas
-            ])
-
+        samples = [ising_sampler(t, np.zeros(n_dim_obs), n=n_samples,
+                    responses=[-1,1]) for t in thetas]
+        data = np.array(samples)
         if time_on_axis == "last":
             data = data.transpose(1, 2, 0)
-        return Bunch(
-            data=data, thetas=np.array(thetas))
+        return data, thetas
 
 def make_dataset(
         n_samples=100, n_dim_obs=100, n_dim_lat=10, T=10, mode=None,
@@ -153,6 +147,8 @@ def make_dataset(
                               proportional=proportional, **kwargs)
 
     elif distribution.lower() == 'ising':
-        return _ising_case(_ising_case(n_samples=100, n_dim_obs=100, T=10,
-                                time_on_axis='first',update_theta='l2',
-                                **kwargs))
+        print(update_theta)
+        return _ising_case(n_samples=n_samples, n_dim_obs=n_dim_obs,
+                                        T=T, time_on_axis=time_on_axis,
+                                update_theta=update_theta,
+                                responses=[-1,1])
